+++ bpk-lib/crates/core/tests/read_walk_evidence_report.rs
@@ -129,7 +129,10 @@ fn read_walk_visibility_matches_plain_query_across_hidden_gap() -> TestResult {
         plain_ids, evidence_ids,
         "PROPERTY: read-walk evidence must apply the same hidden-range visibility predicate as plain query"
     );
-    assert_eq!(plain_ids, vec![before_gap.event_id, after_gap.event_id]);
+    assert_eq!(
+        plain_ids,
+        vec![u128::from(before_gap.event_id), u128::from(after_gap.event_id)]
+    );
     assert_eq!(report.body.matched_count, 2);
     assert_eq!(report.body.returned_count, 2);
     assert!(
